+++ R/count.R
@@ -18,6 +18,14 @@ process_summaries.count_layer <- function(x, ...) {
                    as_label(where),
                    "` is invalid. Filter error:\n", e))
     })
+
+    # Rename missing values
+    for(i in seq_along(missing_count_list)) {
+      built_target <- built_target %>%
+        mutate(!!target_var[[1]] := ifelse(!!target_var[[1]] %in% missing_count_list[[i]], names(missing_count_list)[i], !!target_var[[1]]))
+
+    }
+
   }, envir = x)
 
 
@@ -68,13 +76,23 @@ process_single_count_target <- function(x) {
     if(is.null(include_total_row)) include_total_row <- FALSE
     if(is.null(total_row_label)) total_row_label <- "Total"
 
-
     # The current environment should be the layer itself
     process_count_n(current_env())
 
     if(!is.null(distinct_by)) process_count_distinct_n(current_env())
 
-    if(include_total_row) process_count_total_row(current_env())
+    if(include_total_row){
+      process_count_total_row(current_env())
+      if(!is.null(distinct_by)) {
+        process_count_distinct_total_row(current_env())
+      }
+
+      if(!total_denom_ignore && !is.null(total_count_format) && !is.null(denom_ignore) &&
+         ("pct" %in% total_count_format$vars || "distinct_pct" %in% total_count_format$vars)) {
+           warning("Your total row is ignoring certain values. The 'pct' in this row may not be 100%",
+                   immediate. = TRUE)
+         }
+    }
 
     if(is.null(count_row_prefix)) count_row_prefix <- ""
 
@@ -90,9 +108,16 @@ process_single_count_target <- function(x) {
       ungroup()
 
 
-    if(!is.null(distinct_stat)) numeric_data <- bind_cols(numeric_data,
-                                                          distinct_stat[, c("distinct_n", "distinct_total")])
-
+    if(!is.null(distinct_stat)) {
+      if(include_total_row) {
+        distinct_stat <- distinct_stat %>%
+          bind_rows(total_stat_denom) %>%
+          group_by(!!!denoms_by) %>%
+          do(get_denom_total(., denoms_by, denoms_df, denoms_distinct_df, "distinct_n"))
+      }
+      numeric_data <- bind_cols(numeric_data,
+                                distinct_stat[, c("distinct_n", "distinct_total")])
+    }
   }, envir = x)
 }
 
@@ -175,13 +200,13 @@ process_count_n <- function(x) {
     # If there is a missing_count_string, but its not in the dataset
     if(!is.null(missing_count_string) &&
 
-       !((missing_string %in% unique(built_target[, as_name(target_var[[1]])])) ||
+       !((unname(unlist(missing_count_list)) %in% unique(built_target[, as_name(target_var[[1]])])) ||
         any(is.na(built_target[, as_name(target_var[[1]])])))) {
       # This adds the missing string as a factor to the tallies. This is needed
       # to make sure the missing row is added even if there are no missing values.
       summary_stat <- summary_stat %>%
-        mutate(!!target_var[[1]] := fct_expand(.data[[as_name(target_var[[1]])]],
-                                               missing_string))
+        mutate(!!target_var[[1]] := fct_expand(as.character(.data[[as_name(target_var[[1]])]]),
+                                               names(missing_count_list)))
     }
 
     summary_stat <- summary_stat %>%
@@ -226,15 +251,17 @@ process_count_distinct_n <- function(x) {
       tally(name = "distinct_n") %>%
       ungroup()
 
-      if(!is.null(missing_count_string) &&
-         !(missing_string %in% unique(built_target[, as_name(target_var[[1]])]) ||
-           any(is.na(built_target[, as_name(target_var[[1]])])))) {
-        # This adds the missing string as a factor to the tallies. This is needed
-        # to make sure the missing row is added even if there are no missing values.
-        distinct_stat <- distinct_stat %>%
-          mutate(!!target_var[[1]] := fct_expand(.data[[as_name(target_var[[1]])]],
-                                                 missing_string))
-      }
+    if(!is.null(missing_count_string) &&
+
+       !((unname(unlist(missing_count_list)) %in% unique(built_target[, as_name(target_var[[1]])])) ||
+         any(is.na(built_target[, as_name(target_var[[1]])])))) {
+      # This adds the missing string as a factor to the tallies. This is needed
+      # to make sure the missing row is added even if there are no missing values.
+      summary_stat <- summary_stat %>%
+        mutate(!!target_var[[1]] := fct_expand(as.character(.data[[as_name(target_var[[1]])]]),
+                                               names(missing_count_list)))
+    }
+
 
     # complete all combinations of factors to include combinations that don't exist.
     # add 0 for combinations that don't exist
@@ -280,9 +307,18 @@ change this behavior, use `set_denoms_by()`.", immediate. = TRUE)
           as_name(x) != map_chr(cols, as_name))
     }, treat_var, cols)
 
+    #Create an expression to evaluate filter
+    if(total_denom_ignore){
+      filter_logic <- expr(!(!!target_var[[1]] %in% unlist(denom_ignore)))
+    } else {
+      filter_logic <- expr(TRUE)
+    }
 
     # create a data.frame to create total counts
     total_stat <- summary_stat %>%
+      #Filter out any ignored denoms
+      filter(!!filter_logic) %>%
+      # Use distinct if this is a distinct total row
       # Group by all column variables
       group_by(!!treat_var, !!!cols, !!!denoms_by[needed_denoms_by]) %>%
       summarize(n = sum(n)) %>%
@@ -297,6 +333,55 @@ change this behavior, use `set_denoms_by()`.", immediate. = TRUE)
       ungroup() %>%
       # complete based on missing groupings
       complete(!!treat_var, !!!cols, fill = list(n = 0, total = 0))
+  }, envir = x)
+}
+
+process_count_distinct_total_row <- function(x) {
+  evalq({
+
+    # Check if denoms_by wasn't passed and by was passed.
+    if(is.null(denoms_by) & any(map_lgl(by, quo_is_symbol)) > 0) {
+      warning("A total row was added in addition to non-text by variables, but
+no denoms_by variable was set. This may cause unexpected results. If you wish to
+change this behavior, use `set_denoms_by()`.", immediate. = TRUE)
+    }
+
+    # Make sure the denoms_by is stripped
+    # Stripped of cols and treat_var variables, otherwise it will error out in the group_by
+    # I thought of replacing the group by with !!!unique(c(treat_var, cols, denoms_by))
+    # but that doesn't work due to the denoms_by having an environment set
+
+    # Logical vector that is used to remove the treat_var and cols
+    needed_denoms_by <- map_lgl(denoms_by, function(x, treat_var, cols) {
+      all(as_name(x) != as_name(treat_var),
+          as_name(x) != map_chr(cols, as_name))
+    }, treat_var, cols)
+
+    #Create an expression to evaluate filter
+    if(total_denom_ignore){
+      filter_logic <- expr(!(!!target_var[[1]] %in% unlist(denom_ignore)))
+    } else {
+      filter_logic <- expr(TRUE)
+    }
+
+    # create a data.frame to create total counts
+    total_stat_denom <- distinct_stat %>%
+      #Filter out any ignored denoms
+      filter(!!filter_logic) %>%
+      # Group by all column variables
+      group_by(!!treat_var, !!!cols, !!!denoms_by[needed_denoms_by]) %>%
+      summarize(distinct_n = sum(distinct_n)) %>%
+      ungroup() %>%
+      mutate(distinct_total = distinct_n) %>%
+      # Create a variable to label the totals when it is merged in.
+      mutate(!!as_name(target_var[[1]]) := total_row_label) %>%
+      # Create variables to carry forward 'by'. Only pull out the ones that
+      # aren't symbols
+      group_by(!!!extract_character_from_quo(by)) %>%
+      # ungroup right away to make sure the complete works
+      ungroup() %>%
+      # complete based on missing groupings
+      complete(!!treat_var, !!!cols, fill = list(distinct_n = 0, distinct_total = 0))
   }, envir = x)
 }
 
@@ -364,7 +449,11 @@ process_formatting.count_layer <- function(x, ...) {
     #used to split the string.
     indentation_length <- ifelse(is.null(indentation), 0, nchar(encodeString(indentation)))
 
-    formatted_data <- numeric_data %>%
+    formatted_data <- numeric_data
+
+
+
+    formatted_data <- formatted_data %>%
       # Mutate value based on if there is a distinct_by
       mutate(n = {
         construct_count_string(.n=n, .total=total,
@@ -375,19 +464,15 @@ process_formatting.count_layer <- function(x, ...) {
                                missing_string = missing_string,
                                missing_f_str = missing_count_string,
                                summary_var = summary_var,
-                               indentation_length = indentation_length)
+                               indentation_length = indentation_length,
+                               total_count_format = total_count_format,
+                               total_row_label = total_row_label)
       }) %>%
-
-      # Rename missing values
-      mutate(summary_var = ifelse(summary_var %in% missing_string, missing_name, summary_var)) %>%
-
       # Pivot table
-      pivot_wider(id_cols = c(match_exact(by), "summary_var"),
+        pivot_wider(id_cols = c(match_exact(by), "summary_var"),
                   names_from = c(!!treat_var, match_exact(cols)), values_from = n,
-                  names_prefix = "var1_")
-
+                  names_prefix = "var1_") %>%
     # Replace the by variables and target variable names with `row_label<n>`
-    formatted_data <- formatted_data %>%
       replace_by_string_names(quos(!!!by, summary_var))
 
     if(is_built_nest) {
@@ -448,12 +533,15 @@ process_formatting.count_layer <- function(x, ...) {
 #' @noRd
 construct_count_string <- function(.n, .total, .distinct_n = NULL, .distinct_total = NULL,
                                    count_fmt = NULL, max_layer_length, max_n_width, missing_string,
-                                   missing_f_str, summary_var, indentation_length) {
+                                   missing_f_str, summary_var, indentation_length, total_count_format,
+                                   total_row_label) {
 
   ## Added this for processing formatting in nested count layers where this won't be processed yet
   if (is.null(max_layer_length)) max_layer_length <- 0
   if (is.null(max_n_width)) max_n_width <- 0
   missing_rows <- FALSE
+  total_rows <- FALSE
+
   if (!is.null(missing_f_str)) {
 
     # This subsets the indentation length for nested count layers. The 'outer'
@@ -465,6 +553,12 @@ construct_count_string <- function(.n, .total, .distinct_n = NULL, .distinct_tot
     missing_vars_ord <- map_chr(missing_f_str$vars, as_name)
   }
 
+  ## Pull out string information for total rows
+  if(!is.null(total_count_format)) {
+    total_rows <- summary_var %in% total_row_label
+    total_vars_ord <- map_chr(total_count_format$vars, as_name)
+  }
+
   vars_ord <- map_chr(count_fmt$vars, as_name)
 
   # str_all is a list that contains character vectors for each parameter that might be calculated
@@ -473,37 +567,43 @@ construct_count_string <- function(.n, .total, .distinct_n = NULL, .distinct_tot
   str_all[1] <- count_fmt$repl_str
   # Iterate over every variable
   for(i in seq_along(vars_ord)) {
-    str_all[[i+1]] <-  count_string_switch_help(vars_ord[i], count_fmt, .n[!missing_rows], .total[!missing_rows],
-                                                .distinct_n[!missing_rows], .distinct_total[!missing_rows], vars_ord)
+    str_all[[i+1]] <-  count_string_switch_help(vars_ord[i], count_fmt, .n[!missing_rows & !total_rows], .total[!missing_rows & !total_rows],
+                                                .distinct_n[!missing_rows & !total_rows], .distinct_total[!missing_rows & !total_rows], vars_ord)
   }
 
 
   # Logic for missing
-  if(!is.null(missing_f_str)) {
-    # Same logic as above, just add for missing
-    missing_str_all <- vector("list", 5)
-    missing_str_all[1] <- missing_f_str$repl_str
-    for(i in seq_along(missing_vars_ord)) {
-      missing_str_all[[i+1]] <- count_string_switch_help(missing_vars_ord[i], missing_f_str, .n[missing_rows], .total[missing_rows],
-                                                         .distinct_n[missing_rows], .distinct_total[missing_rows], missing_vars_ord)
-    }
-
-    # Put the vector strings together. Only include parts of str_all that aren't null
-    string_nm <- do.call(sprintf, str_all[!map_lgl(str_all, is.null)])
-    string_m <- do.call(sprintf, missing_str_all[!map_lgl(missing_str_all, is.null)])
-    string_ <- character(length(string_nm) + length(string_m))
-    string_[!missing_rows] <- string_nm
-    string_[missing_rows] <- string_m
-
-    # If there is no missing
-  } else {
-
-    string_ <- do.call(sprintf, str_all[!map_lgl(str_all, is.null)])
-
+  # Same logic as above, just add for missing
+  missing_str_all <- vector("list", 5)
+  missing_str_all[1] <- missing_f_str$repl_str
+  for(i in seq_along(missing_vars_ord)) {
+    missing_str_all[[i+1]] <- count_string_switch_help(missing_vars_ord[i], missing_f_str, .n[missing_rows], .total[missing_rows],
+                                                       .distinct_n[missing_rows], .distinct_total[missing_rows], missing_vars_ord)
   }
 
+  total_str_all <- vector("list", 5)
+  total_str_all[1] <- total_count_format$repl_str
+  for(i in seq_along(total_vars_ord)) {
+    total_str_all[[i+1]] <- count_string_switch_help(total_vars_ord[i], total_count_format, .n[total_rows], .total[total_rows],
+                                                     .distinct_n[total_rows], .distinct_total[total_rows], total_vars_ord)
+  }
 
-  # Left pad set to 0 so it won't pad to the left anymore.
+  # Put the vector strings together. Only include parts of str_all that aren't null
+  # nm is non-missing, m is mising, and t is total.
+  string_nm <- do.call(sprintf, str_all[!map_lgl(str_all, is.null)])
+  if(!is.null(missing_vars_ord)) string_m <-  do.call(sprintf, missing_str_all[!map_lgl(missing_str_all, is.null)])
+  if(!is.null(total_vars_ord)) string_t <- do.call(sprintf, total_str_all[!map_lgl(total_str_all, is.null)])
+  # string_ is the final string to return. Merge the missing, non-missing, and others together
+  string_ <- character(length(string_nm) + length(string_m) + length(string_t))
+  string_[!missing_rows & !total_rows] <- string_nm
+  string_[total_rows] <-   string_t
+  string_[missing_rows] <-  string_m
+
+
+
+
+  # Left pad set to 0 meaning it won't pad to the left at all
+  # right pad is set to the maximum n count in the table
   string_ <- pad_formatted_data(string_, 0, max_n_width)
 
   string_
@@ -582,6 +682,11 @@ process_count_denoms <- function(x) {
     param_apears <- map_lgl(layer_params, function(x) {
       as_name(x) %in% names(target)
     })
+
+    # Raise errors if a denom is ignored but there isn't a missing count string
+    if(!is.null(denom_ignore) && is.null(missing_count_string)) {
+      abort("A value(s) were set with 'denom_ignore' but no missing count was set. Your percentages/totals may not have meaning.")
+    }
 
     # Subset the local built_target based on where
     # Catch errors